+++ src/pymodaq_plugins_hamamatsu/daq_viewer_plugins/plugins_1D/daq_1Dviewer_MiniSpectro.py
@@ -4,7 +4,10 @@ from pymodaq.utils.data import DataFromPlugins, Axis, DataToExport
 from pymodaq.control_modules.viewer_utility_classes import DAQ_Viewer_base, comon_parameters, main
 from pymodaq.utils.parameter import Parameter
 
-from pymodaq_plugins_hamamatsu.hardware.minispectro import MiniSpectro
+
+class PythonWrapperOfYourInstrument:
+    #  TODO Replace this fake class with the import of the real python wrapper of your instrument
+    pass
 
 # TODO:
 # (1) change the name of the following class to DAQ_1DViewer_TheNameOfYourChoice
@@ -12,13 +15,11 @@ from pymodaq_plugins_hamamatsu.hardware.minispectro import MiniSpectro
 #     for the class name and the file name.)
 # (3) this file should then be put into the right folder, namely IN THE FOLDER OF THE PLUGIN YOU ARE DEVELOPING:
 #     pymodaq_plugins_my_plugin/daq_viewer_plugins/plugins_1D
-class DAQ_1DViewer_MiniSpectro(DAQ_Viewer_base):
+class DAQ_1DViewer_Template(DAQ_Viewer_base):
     """ Instrument plugin class for a 1D viewer.
     
     This object inherits all functionalities to communicate with PyMoDAQ's DAQ_Viewer module through inheritance via
     DAQ_Viewer_base. It makes a bridge between the DAQ_Viewer module and the Python wrapper of a particular instrument.
-
-    C10083CA USB Hamamatsu mini-spectrometer
 
     TODO Complete the docstring of your plugin with:
         * The set of instruments that should be compatible with this instrument plugin.
@@ -45,7 +46,7 @@ class DAQ_1DViewer_MiniSpectro(DAQ_Viewer_base):
     def ini_attributes(self):
         #  TODO declare the type of the wrapper (and assign it to self.controller) you're going to use for easy
         #  autocompletion
-        self.controller: MiniSpectro = None
+        self.controller: PythonWrapperOfYourInstrument = None
 
         # TODO declare here attributes you want/need to init with a default value
 
@@ -80,22 +81,26 @@ class DAQ_1DViewer_MiniSpectro(DAQ_Viewer_base):
         initialized: bool
             False if initialization failed otherwise True
         """
+
+        raise NotImplemented  # TODO when writing your own plugin remove this line and modify the one below
         self.ini_detector_init(slave_controller=controller)
 
         if self.is_master:
-            self.controller = MiniSpectro(device_name='tm_ccd')  #instantiate you driver with whatever arguments are needed
+            self.controller = PythonWrapperOfYourInstrument()  #instantiate you driver with whatever arguments are needed
+            self.controller.open_communication() # call eventual methods
 
         ## TODO for your custom plugin
         # get the x_axis (you may want to to this also in the commit settings if x_axis may have changed
-        data_x_axis = self.controller.get_sensor_data()[0]  # if possible
-        self.x_axis = Axis(data=data_x_axis, label='Pixels', units='', index=0)
+        data_x_axis = self.controller.your_method_to_get_the_x_axis()  # if possible
+        self.x_axis = Axis(data=data_x_axis, label='', units='', index=0)
 
-        # TODO for your custom plugin. Initialize viewers panel with the future type of data
-        # self.dte_signal_temp.emit(DataToExport(name='MiniSpectro',
-        #                                        data=[DataFromPlugins(name='Mini-spectrometer',
-        #                                                              data=[np.array([10 for _ in range(2048)])],
-        #                                                              dim='Data1D', labels=['Intensity'],
-        #                                                              axes=[self.x_axis])]))
+        # TODO for your custom plugin. Initialize viewers pannel with the future type of data
+        self.dte_signal_temp.emit(DataToExport(name='myplugin',
+                                               data=[DataFromPlugins(name='Mock1',
+                                                                     data=[np.array([0., 0., ...]),
+                                                                           np.array([0., 0., ...])],
+                                                                     dim='Data1D', labels=['Mock1', 'label2'],
+                                                                     axes=[self.x_axis])]))
 
         info = "Whatever info you want to log"
         initialized = True
@@ -103,7 +108,9 @@ class DAQ_1DViewer_MiniSpectro(DAQ_Viewer_base):
 
     def close(self):
         """Terminate the communication protocol"""
-        self.controller.close()
+        ## TODO for your custom plugin
+        raise NotImplemented  # when writing your own plugin remove this line
+        #  self.controller.your_method_to_terminate_the_communication()  # when writing your own plugin replace this line
 
     def grab_data(self, Naverage=1, **kwargs):
         """Start a grab from the detector
@@ -116,20 +123,32 @@ class DAQ_1DViewer_MiniSpectro(DAQ_Viewer_base):
         kwargs: dict
             others optionals arguments
         """
-        # Synchrone version (blocking function)
-        data_tot = self.controller.get_sensor_data()[1]
-        self.dte_signal.emit(DataToExport(name='MiniSpectro',
-                                          data=[DataFromPlugins(name='Mini-spectrometer',
-                                                                data=data_tot,
-                                                                dim='Data1D',
-                                                                labels=['Spectro'],
+        ## TODO for your custom plugin: you should choose EITHER the synchrone or the asynchrone version following
+
+        ##synchrone version (blocking function)
+        data_tot = self.controller.your_method_to_start_a_grab_snap()
+        self.dte_signal.emit(DataToExport('myplugin',
+                                          data=[DataFromPlugins(name='Mock1', data=data_tot,
+                                                                dim='Data1D', labels=['dat0', 'data1'],
                                                                 axes=[self.x_axis])]))
 
-    
+        ##asynchrone version (non-blocking function with callback)
+        self.controller.your_method_to_start_a_grab_snap(self.callback)
+        #########################################################
+
+
+    def callback(self):
+        """optional asynchrone method called when the detector has finished its acquisition of data"""
+        data_tot = self.controller.your_method_to_get_data_from_buffer()
+        self.dte_signal.emit(DataToExport('myplugin',
+                                          data=[DataFromPlugins(name='Mock1', data=data_tot,
+                                                                dim='Data1D', labels=['dat0', 'data1'])]))
+
     def stop(self):
         """Stop the current grab hardware wise if necessary"""
-
-        self.controller.close()  # when writing your own plugin replace this line
+        ## TODO for your custom plugin
+        raise NotImplemented  # when writing your own plugin remove this line
+        self.controller.your_method_to_stop_acquisition()  # when writing your own plugin replace this line
         self.emit_status(ThreadCommand('Update_Status', ['Some info you want to log']))
         ##############################
         return ''
